refactor context enhancement according to changes in presidio 2.2.27
