+++ hebsafeharbor/common/city_utils.py
@@ -1053,7 +1053,6 @@ BELOW_THRESHOLD_CITIES_LIST = ['אבו סריחאן',
  'קוממיות',
  'קורנית',
  'קטורה',
- 'קיבוץ',
  'קלחים',
  'קליה',
  'קרית יערים',
