+++ hebsafeharbor/common/context_terms_recognizer.py
@@ -35,10 +35,10 @@ class ContextTermsRecognizer(TermsRecognizer):
                 found_elements = re.search(pattern, text)
 
                 if found_elements:
-                    supported_word_position = found_elements[2]
-                    supported_word = text[supported_word_position[0]:supported_word_position[1]]
-                    preposition_position = found_elements[3] if prefixes else None
-                    preposition = text[preposition_position[0]:preposition_position[1]] if prefixes else None
+                    supported_word_position = found_elements.regs[2]
+                    supported_word = found_elements[2]
+                    preposition_position = found_elements.regs[3] if prefixes else None
+                    preposition = found_elements[3] if prefixes else None
                     context.append((supported_word, supported_word_position, preposition, preposition_position))
 
         # drop duplicates
