--- hebsafeharbor/identifier/context_enhancement/hebrew_context_aware_enhancer.py
@@ -1,388 +1,0 @@
-import copy
-import logging
-import re
-from typing import List, Optional
-
-from presidio_analyzer import RecognizerResult, LemmaContextAwareEnhancer
-from presidio_analyzer import EntityRecognizer
-from presidio_analyzer.nlp_engine import NlpArtifacts
-from presidio_analyzer.context_aware_enhancers import ContextAwareEnhancer
-
-logger = logging.getLogger("presidio-analyzer")
-
-
-class HebrewContextAwareEnhancer(ContextAwareEnhancer):
-    """
-    A class representing a custom context aware enhancer logic.
-
-    Context words might enhance confidence score of a recognized entity,
-    HebrewContextAwareEnhancer is a customized context aware logic,
-    it enhances specified recognizers outcomes according to custom logic
-    (ex. strict match for context word, additional logic for preposition handling),
-    and duplicates LemmaContextAwareEnhancer for all other classes ,
-    if matched it enhance the recognized entity confidence score by a given factor.
-
-    :param substring_processed_recognizers_list: Results of which recognizers should be enhanced using custom logic
-    :param context_similarity_factor: How much to enhance confidence of match entity
-    :param min_score_with_context_similarity: Minimum confidence score
-    :param context_prefix_count: how many words before the entity to match context
-    :param context_suffix_count: how many words after the entity to match context
-    """
-
-    def __init__(
-            self,
-            substring_processed_recognizers_list: List[str] = None,
-            context_similarity_factor: float = 0.35,
-            min_score_with_context_similarity: float = 0.4,
-            context_prefix_count: int = 5,
-            context_suffix_count: int = 0,
-    ):
-        super().__init__(
-            context_similarity_factor=context_similarity_factor,
-            min_score_with_context_similarity=min_score_with_context_similarity,
-            context_prefix_count=context_prefix_count,
-            context_suffix_count=context_suffix_count,
-        )
-        self.substring_processed_recognizers_list = substring_processed_recognizers_list
-
-    def enhance_using_context(
-            self,
-            text: str,
-            raw_results: List[RecognizerResult],
-            nlp_artifacts: NlpArtifacts,
-            recognizers: List[EntityRecognizer],
-            context: Optional[List[str]] = None,
-    ) -> List[RecognizerResult]:
-        """
-        Update results in case the surrounding words are identical to the context words.
-
-        Using the surrounding words of the actual word matches, look
-        for specific strings that if found contribute to the score
-        of the result, improving the confidence that the match is
-        indeed of that PII entity type
-
-        :param text: The actual text that was analyzed
-        :param raw_results: Recognizer results which didn't take
-                            context into consideration
-        :param nlp_artifacts: The nlp artifacts contains elements
-                              such as lemmatized tokens for better
-                              accuracy of the context enhancement process
-        :param recognizers: the list of recognizers
-        :param context: list of context words
-        """  # noqa D205 D400
-
-        # create a deep copy of the results object so we can manipulate it
-        results = copy.deepcopy(raw_results)
-
-        # create recongnizer context dictionary
-        recognizers_dict = {recognizer.name: recognizer for recognizer in recognizers}
-
-        # Create empty list in None or lowercase all context words in the list
-        if not context:
-            context = []
-        else:
-            context = [word.lower() for word in context]
-
-        # Sanity
-        if nlp_artifacts is None:
-            logger.warning("NLP artifacts were not provided")
-            return results
-
-        for result in results:
-            recognizer = None
-            # get recognizer matching the result, if found.
-            if (
-                    result.recognition_metadata
-                    and RecognizerResult.RECOGNIZER_NAME_KEY
-                    in result.recognition_metadata.keys()
-            ):
-                recognizer = recognizers_dict.get(
-                    result.recognition_metadata[RecognizerResult.RECOGNIZER_NAME_KEY]
-                )
-
-            if not recognizer:
-                logger.debug(
-                    "Recognizer name not found as part of the "
-                    "recognition_metadata dict in the RecognizerResult. "
-                )
-                continue
-
-            # skip recognizer result if the recognizer doesn't support
-            # context enhancement
-            if not recognizer.context:
-                logger.debug(
-                    "recognizer '%s' does not support context enhancement",
-                    recognizer.name,
-                )
-                continue
-
-            # extract lemmatized context from the surrounding of the match
-            word = text[result.start:result.end]
-
-            if recognizer.name in self.substring_processed_recognizers_list:
-                sentence = self.__extract_sentence_containing_position(
-                    nlp_artifacts=nlp_artifacts, start=result.start
-                )
-
-                if sentence != "":
-                    supportive_context = self.__find_supportive_context_in_sentence(
-                        sentence, word, recognizer
-                    )
-                    if supportive_context:
-                        result.score += self.context_similarity_factor
-                        result.score = max(result.score, self.min_score_with_context_similarity)
-                        result.score = min(result.score, ContextAwareEnhancer.MAX_SCORE)
-
-                        # Update the explainability object with context information
-                        # helped improving the score
-                        result.analysis_explanation.set_supportive_context_word(
-                            supportive_context
-                        )
-                        result.analysis_explanation.set_improved_score(result.score)
-
-                elif recognizer.name not in self.substring_processed_recognizers_list:
-                    surrounding_words = self.__extract_surrounding_words(
-                        nlp_artifacts=nlp_artifacts, word=word, start=result.start
-                    )
-
-                    # combine other sources of context with surrounding words
-                    surrounding_words.extend(context)
-
-                    supportive_context_word = self.__find_supportive_word_in_context(
-                        surrounding_words, recognizer.context
-                    )
-                    if supportive_context_word != "":
-                        result.score += self.context_similarity_factor
-                        result.score = max(result.score, self.min_score_with_context_similarity)
-                        result.score = min(result.score, ContextAwareEnhancer.MAX_SCORE)
-
-                        # Update the explainability object with context information
-                        # helped improving the score
-                        result.analysis_explanation.set_supportive_context_word(
-                            supportive_context_word
-                        )
-                        result.analysis_explanation.set_improved_score(result.score)
-
-        return results
-
-    @staticmethod
-    def __find_supportive_word_in_context(
-            context_list: List[str], recognizer_context_list: List[str]
-    ) -> str:
-        """
-        Find words in the text which are relevant for context evaluation.
-
-        A word is considered a supportive context word if there's exact match
-        between a keyword in context_text and any keyword in context_list.
-
-        :param context_list words before and after the matched entity within
-               a specified window size
-        :param recognizer_context_list a list of words considered as
-                context keywords manually specified by the recognizer's author
-        """
-        word = ""
-        # If the context list is empty, no need to continue
-        if context_list is None or recognizer_context_list is None:
-            return word
-
-        for predefined_context_word in recognizer_context_list:
-            # result == true only if any of the predefined context words
-            # is found exactly or as a substring in any of the collected
-            # context words
-            result = next(
-                (
-                    True
-                    for keyword in context_list
-                    if predefined_context_word in keyword
-                ),
-                False,
-            )
-            if result:
-                logger.debug("Found context keyword '%s'", predefined_context_word)
-                word = predefined_context_word
-                break
-
-        return word
-
-    def __extract_surrounding_words(
-            self, nlp_artifacts: NlpArtifacts, word: str, start: int
-    ) -> List[str]:
-        """Extract words surrounding another given word.
-
-        The text from which the context is extracted is given in the nlp
-        doc.
-
-        :param nlp_artifacts: An abstraction layer which holds different
-                              items which are the result of a NLP pipeline
-                              execution on a given text
-        :param word: The word to look for context around
-        :param start: The start index of the word in the original text
-        """
-        if not nlp_artifacts.tokens:
-            logger.info("Skipping context extraction due to lack of NLP artifacts")
-            # if there are no nlp artifacts, this is ok, we can
-            # extract context and we return a valid, yet empty
-            # context
-            return [""]
-
-        # Get the already prepared words in the given text, in their
-        # LEMMATIZED version
-        lemmatized_keywords = nlp_artifacts.keywords
-
-        # since the list of tokens is not necessarily aligned
-        # with the actual index of the match, we look for the
-        # token index which corresponds to the match
-        token_index = self._find_index_of_match_token(
-            word, start, nlp_artifacts.tokens, nlp_artifacts.tokens_indices
-        )
-
-        # index i belongs to the PII entity, take the preceding n words
-        # and the successing m words into a context list
-
-        backward_context = self.__add_n_words_backward(
-            token_index,
-            self.context_prefix_count,
-            nlp_artifacts.lemmas,
-            lemmatized_keywords,
-        )
-        forward_context = self.__add_n_words_forward(
-            token_index,
-            self.context_suffix_count,
-            nlp_artifacts.lemmas,
-            lemmatized_keywords,
-        )
-
-        context_list = []
-        context_list.extend(backward_context)
-        context_list.extend(forward_context)
-        context_list = list(set(context_list))
-        logger.debug("Context list is: %s", " ".join(context_list))
-        return context_list
-
-    @staticmethod
-    def _find_index_of_match_token(
-            word: str, start: int, tokens, tokens_indices: List[int]  # noqa ANN001
-    ) -> int:
-        found = False
-        # we use the known start index of the original word to find the actual
-        # token at that index, we are not checking for equivilance since the
-        # token might be just a substring of that word (e.g. for phone number
-        # 555-124564 the first token might be just '555' or for a match like '
-        # rocket' the actual token will just be 'rocket' hence the misalignment
-        # of indices)
-        # Note: we are iterating over the original tokens (not the lemmatized)
-        i = -1
-        for i, token in enumerate(tokens, 0):
-            # Either we found a token with the exact location, or
-            # we take a token which its characters indices covers
-            # the index we are looking for.
-            if (tokens_indices[i] == start) or (start < tokens_indices[i] + len(token)):
-                # found the interesting token, the one that around it
-                # we take n words, we save the matching lemma
-                found = True
-                break
-
-        if not found:
-            raise ValueError(
-                "Did not find word '" + word + "' "
-                                               "in the list of tokens although it "
-                                               "is expected to be found"
-            )
-        return i
-
-    @staticmethod
-    def __add_n_words(
-            index: int,
-            n_words: int,
-            lemmas: List[str],
-            lemmatized_filtered_keywords: List[str],
-            is_backward: bool,
-    ) -> List[str]:
-        """
-        Prepare a string of context words.
-
-        Return a list of words which surrounds a lemma at a given index.
-        The words will be collected only if exist in the filtered array
-
-        :param index: index of the lemma that its surrounding words we want
-        :param n_words: number of words to take
-        :param lemmas: array of lemmas
-        :param lemmatized_filtered_keywords: the array of filtered
-               lemmas from the original sentence,
-        :param is_backward: if true take the preceeding words, if false,
-                            take the successing words
-        """
-        i = index
-        context_words = []
-        # The entity itself is no interest to us...however we want to
-        # consider it anyway for cases were it is attached with no spaces
-        # to an interesting context word, so we allow it and add 1 to
-        # the number of collected words
-
-        # collect at most n words (in lower case)
-        remaining = n_words + 1
-        while 0 <= i < len(lemmas) and remaining > 0:
-            lower_lemma = lemmas[i].lower()
-            if lower_lemma in lemmatized_filtered_keywords:
-                context_words.append(lower_lemma)
-                remaining -= 1
-            i = i - 1 if is_backward else i + 1
-        return context_words
-
-    def __add_n_words_forward(
-            self,
-            index: int,
-            n_words: int,
-            lemmas: List[str],
-            lemmatized_filtered_keywords: List[str],
-    ) -> List[str]:
-        return self.__add_n_words(
-            index, n_words, lemmas, lemmatized_filtered_keywords, False
-        )
-
-    def __add_n_words_backward(
-            self,
-            index: int,
-            n_words: int,
-            lemmas: List[str],
-            lemmatized_filtered_keywords: List[str],
-    ) -> List[str]:
-        return self.__add_n_words(
-            index, n_words, lemmas, lemmatized_filtered_keywords, True
-        )
-
-    @staticmethod
-    def __extract_sentence_containing_position(nlp_artifacts: NlpArtifacts, start: int) -> str:
-        """
-        Extracting the sentence which contains the start position of recognized entity using sentences from nlp_artifacts
-        :param nlp_artifacts: An abstraction layer which holds different
-                              items which are the result of a NLP pipeline
-                              execution on a given text
-        :param start: The start index of the word in the original text
-        :return: sentence string
-        """
-        for sent in nlp_artifacts.tokens.sents:
-            if (start >= sent.start_char) & (start <= sent.end_char):
-                return sent.text
-        return ''
-
-    def __find_supportive_context_in_sentence(self, sentence: str, word: str, recognizer: EntityRecognizer):
-        """
-        Extracting the sentence which contains the start position of recognized entity using sentences from nlp_artifacts
-        :param sentence: sentence string
-        :param word: recognized entity string
-        :param recognizer: entity recognizer
-        :return: details of supportive context recognized
-        """
-        # Sanity
-        if not recognizer.context_recognizer:
-            logger.warning("Context recognizer is not defined")
-            return None
-        if recognizer.allowed_prepositions:
-            prep = recognizer.allowed_prepositions
-        else:
-            logger.warning("Prepositions are not defined")
-            prep = None
-
-        contexts = recognizer.context_recognizer(sentence, word, prep)
-        return contexts
